git-svn-id: http://127.0.0.1/svn/trunk@312 02e793a8-47cb-4219-9cc1-7115cfedab66

+++ src/starcorp/server/entitystore/HibernateStore.java
@@ -655,7 +655,7 @@ public class HibernateStore implements IEntityStore {
 		return entity;
 	}
 
-	public ABaseEntity load(Class<?> entityClass, int ID) {
+	public ABaseEntity load(Class<?> entityClass, long ID) {
 		return (ABaseEntity) loadObject(entityClass, ID);
 	}
